requirements.txt
beautifulsoup4==4.11.1
grpcio==1.47.0
grpcio-tools==1.47.0
six==1.16.0
soupsieve==2.3.2.post1
